refactor(chat): Split Chat component into sub-components
